Feat(Suspend): add file to declare js functions that are common used in all doctypes to reduce declaring in every doctype and delete them from doctype files.
